cleanup capability; remove debug;
